turned permission enforcer into abstract class
