+++ src/app/core/entity/entity-mapper/mock-entity-mapper-service.ts
@@ -20,7 +20,6 @@ export function mockEntityMapperProvider(
   customDatabaseResolverService: DatabaseResolverService | {} = {
     resetSync: () => Promise.resolve(),
     isIndexedDbAdapterSupported: () => false,
-    hasLocalDatabase: () => true,
     getDatabase: () => ({ purge: () => Promise.resolve(false) }),
   },
 ): Provider[] {
