+++ src/app/core/permissions/permission-enforcer/noop-permission-enforcer.service.spec.ts
@@ -1,0 +1,25 @@
+import { NoopPermissionEnforcerService } from "./noop-permission-enforcer.service";
+
+describe("NoopPermissionEnforcerService", () => {
+  let service: NoopPermissionEnforcerService;
+
+  beforeEach(() => {
+    service = new NoopPermissionEnforcerService();
+  });
+
+  it("should be created", () => {
+    expect(service).toBeTruthy();
+  });
+
+  it("should resolve without doing anything", async () => {
+    await expect(
+      service.enforcePermissionsOnLocalData([
+        { subject: "all", action: "manage" },
+      ]),
+    ).resolves.toBeUndefined();
+  });
+
+  it("should return undefined for last enforced rules", () => {
+    expect(service.getLastEnforcedRules()).toBeUndefined();
+  });
+});
